pyproject.toml
[project]
name = "2quip-agent"
version = "0.1.0"
description = "2Quip AI Agent with LiveKit Voice Integration"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "agno>=2.3.24",
    "ddgs>=9.10.0",
    "fastapi[all]>=0.128.0",
    "groq>=1.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy-libsql>=0.2.0",
    # LiveKit Voice Agent dependencies
    "livekit",
    "livekit-agents",
    "livekit-plugins-silero",
    "livekit-plugins-noise-cancellation",
    "livekit-plugins-turn-detector",
    "livekit-plugins-assemblyai",
    "livekit-plugins-cartesia",
]

[project.optional-dependencies]
# Optional TTS/STT providers - install as needed
deepgram = ["livekit-plugins-deepgram"]
cartesia = ["livekit-plugins-cartesia"]
assemblyai = ["livekit-plugins-assemblyai"]
openai-realtime = ["livekit-plugins-openai"]

# All voice providers
voice-all = [
    "livekit-plugins-deepgram",
    "livekit-plugins-cartesia",
    "livekit-plugins-assemblyai",
    "livekit-plugins-openai",
]

[dependency-groups]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "httpx>=0.27"]
